Struct google_dfareporting3d3::api::Campaign

pub struct Campaign {
    pub account_id: Option<String>,
    pub ad_blocking_configuration: Option<AdBlockingConfiguration>,
    pub additional_creative_optimization_configurations: Option<Vec<CreativeOptimizationConfiguration>>,
    pub advertiser_group_id: Option<String>,
    pub advertiser_id: Option<String>,
    pub advertiser_id_dimension_value: Option<DimensionValue>,
    pub archived: Option<bool>,
    pub audience_segment_groups: Option<Vec<AudienceSegmentGroup>>,
    pub billing_invoice_code: Option<String>,
    pub click_through_url_suffix_properties: Option<ClickThroughUrlSuffixProperties>,
    pub comment: Option<String>,
    pub create_info: Option<LastModifiedInfo>,
    pub creative_group_ids: Option<Vec<String>>,
    pub creative_optimization_configuration: Option<CreativeOptimizationConfiguration>,
    pub default_click_through_event_tag_properties: Option<DefaultClickThroughEventTagProperties>,
    pub default_landing_page_id: Option<String>,
    pub end_date: Option<String>,
    pub event_tag_overrides: Option<Vec<EventTagOverride>>,
    pub external_id: Option<String>,
    pub id: Option<String>,
    pub id_dimension_value: Option<DimensionValue>,
    pub kind: Option<String>,
    pub last_modified_info: Option<LastModifiedInfo>,
    pub name: Option<String>,
    pub nielsen_ocr_enabled: Option<bool>,
    pub start_date: Option<String>,
    pub subaccount_id: Option<String>,
    pub trafficker_emails: Option<Vec<String>>,
}

Contains properties of a Campaign Manager campaign.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get campaigns (response)

insert campaigns (request|response)

list campaigns (none)

patch campaigns (request|response)

update campaigns (request|response)

Fields
account_id: Option<String>

Account ID of this campaign. This is a read-only field that can be left blank.

ad_blocking_configuration: Option<AdBlockingConfiguration>

Ad blocking settings for this campaign.

additional_creative_optimization_configurations: Option<Vec<CreativeOptimizationConfiguration>>

Additional creative optimization configurations for the campaign.

advertiser_group_id: Option<String>

Advertiser group ID of the associated advertiser.

advertiser_id: Option<String>

Advertiser ID of this campaign. This is a required field.

advertiser_id_dimension_value: Option<DimensionValue>

Dimension value for the advertiser ID of this campaign. This is a read-only, auto-generated field.

archived: Option<bool>

Whether this campaign has been archived.

audience_segment_groups: Option<Vec<AudienceSegmentGroup>>

Audience segment groups assigned to this campaign. Cannot have more than 300 segment groups.

billing_invoice_code: Option<String>

Billing invoice code included in the Campaign Manager client billing invoices associated with the campaign.

click_through_url_suffix_properties: Option<ClickThroughUrlSuffixProperties>

Click-through URL suffix override properties for this campaign.

comment: Option<String>

Arbitrary comments about this campaign. Must be less than 256 characters long.

create_info: Option<LastModifiedInfo>

Information about the creation of this campaign. This is a read-only field.

creative_group_ids: Option<Vec<String>>

List of creative group IDs that are assigned to the campaign.

creative_optimization_configuration: Option<CreativeOptimizationConfiguration>

Creative optimization configuration for the campaign.

default_click_through_event_tag_properties: Option<DefaultClickThroughEventTagProperties>

Click-through event tag ID override properties for this campaign.

default_landing_page_id: Option<String>

The default landing page ID for this campaign.

end_date: Option<String>

no description provided

event_tag_overrides: Option<Vec<EventTagOverride>>

Overrides that can be used to activate or deactivate advertiser event tags.

external_id: Option<String>

External ID for this campaign.

id: Option<String>

ID of this campaign. This is a read-only auto-generated field.

id_dimension_value: Option<DimensionValue>

Dimension value for the ID of this campaign. This is a read-only, auto-generated field.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#campaign".

last_modified_info: Option<LastModifiedInfo>

Information about the most recent modification of this campaign. This is a read-only field.

name: Option<String>

Name of this campaign. This is a required field and must be less than 256 characters long and unique among campaigns of the same advertiser.

nielsen_ocr_enabled: Option<bool>

Whether Nielsen reports are enabled for this campaign.

start_date: Option<String>

no description provided

subaccount_id: Option<String>

Subaccount ID of this campaign. This is a read-only field that can be left blank.

trafficker_emails: Option<Vec<String>>

Campaign trafficker contact emails.